memidahkan file stack test
